OOP changes and refactored response code
The response codes are now stored in a hash. After attempting to
implement constants for this, it became apparent it's a bad idea since
perl will translate them as barewords (subprocesses).

To ensure that less code is being written, %RESPONSE has been given the
'our' declaration so that all modules within the DovecotBcryptAuth scope
have access to it if they need it. This also prompted the move of all
modules residing inside of '/lib', now residing in
/lib/DovecotBcryptAuth'.